+++ crates/polars-io/src/prelude.rs
@@ -1,6 +1,3 @@
-#[cfg(test)]
-use polars_core::prelude::*;
-
 pub use crate::cloud;
 #[cfg(feature = "csv")]
 pub use crate::csv::{read::*, write::*};
@@ -14,10 +11,3 @@ pub use crate::ndjson::core::*;
 pub use crate::parquet::{metadata::*, read::*, write::*};
 pub use crate::shared::{SerReader, SerWriter};
 pub use crate::utils::*;
-
-#[cfg(test)]
-pub(crate) fn create_df() -> DataFrame {
-    let s0 = Series::new("days", [0, 1, 2, 3, 4].as_ref());
-    let s1 = Series::new("temp", [22.1, 19.9, 7., 2., 3.].as_ref());
-    DataFrame::new(vec![s0, s1]).unwrap()
-}
